requirements.txt
python-telegram-bot
tortoise-orm
asyncpg
dotenv
pillow
google-genai